chore(tests): changed files layout
