+++ ledger/interface.go
